rename sln

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -24,13 +25,13 @@ public class AuctionsController : ControllerBase
         _publishEndpoint = publishEndpoint;
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        // TODO: add current user as a seller
 
-        auction.Seller = "wizard";
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
 
@@ -71,6 +72,7 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuctionById(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -79,6 +81,8 @@ public class AuctionsController : ControllerBase
                                     .FirstOrDefaultAsync(x => x.Id == id);
 
        if (auction == null) return NotFound();
+
+       if (auction.Seller != User.Identity.Name) return Forbid("");
 
        auction.Item.Type = updateAuctionDto.Type ?? auction.Item.Type;
        auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -95,6 +99,7 @@ public class AuctionsController : ControllerBase
        return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuctionById(Guid id)
     {
@@ -102,7 +107,7 @@ public class AuctionsController : ControllerBase
 
         if (auction == null) return NotFound();
 
-        // TO: check seller equal to username
+        if (auction.Seller != User.Identity.Name) return Forbid("");
 
         _context.Auctions.Remove(auction);
